+++ whitenoise/__init__.py
@@ -1,5 +1,5 @@
 from .base import WhiteNoise
 
-__version__ = '3.3.0'
+__version__ = '4.0b1'
 
 __all__ = ['WhiteNoise']